+++ libavcodec/vvc/vvcdsp.h
@@ -152,16 +152,16 @@ typedef struct VVCSAODSPContext {
 
 typedef struct VVCALFDSPContext {
     void (*filter[2 /* luma, chroma */])(uint8_t *dst, ptrdiff_t dst_stride, const uint8_t *src, ptrdiff_t src_stride,
-        int width, int height, const int8_t *filter, const int16_t *clip);
+        int width, int height, const int16_t *filter, const int16_t *clip);
     void (*filter_vb[2 /* luma, chroma */])(uint8_t *dst, ptrdiff_t dst_stride, const uint8_t *src,  ptrdiff_t src_stride,
-        int width, int height, const int8_t *filter, const int16_t *clip, int vb_pos);
+        int width, int height, const int16_t *filter, const int16_t *clip, int vb_pos);
     void (*filter_cc)(uint8_t *dst, ptrdiff_t dst_stride, const uint8_t *luma, ptrdiff_t luma_stride,
-        int width, int height, int hs, int vs, const int8_t *filter, int vb_pos);
+        int width, int height, int hs, int vs, const int16_t *filter, int vb_pos);
 
     void (*classify)(int *class_idx, int *transpose_idx, const uint8_t *src, ptrdiff_t src_stride, int width, int height,
         int vb_pos, int *gradient_tmp);
-    void (*recon_coeff_and_clip)(int8_t *coeff, int16_t *clip, const int *class_idx, const int *transpose_idx, int size,
-        const int8_t *coeff_set, const uint8_t *clip_idx_set, const uint8_t *class_to_filt);
+    void (*recon_coeff_and_clip)(int16_t *coeff, int16_t *clip, const int *class_idx, const int *transpose_idx, int size,
+        const int16_t *coeff_set, const uint8_t *clip_idx_set, const uint8_t *class_to_filt);
 } VVCALFDSPContext;
 
 typedef struct VVCDSPContext {
